requirements.txt
asgiref==3.5.2
autopep8==2.0.0
bleach==5.0.1
crispy-bootstrap5==0.7
Django==4.1.3
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
fontawesomefree==6.2.1
gunicorn==20.1.0
imgkit==1.2.2
Pillow==9.3.0
psycopg2-binary==2.9.1
pycodestyle==2.10.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
webencodings==0.5.1
whitenoise==6.2.0
django-admin-thumbnails==0.2.6
requests==2.28.1

humanize==4.4.0